feat: 레시피 페이지 목록 조회(조회수, 평점 정렬)

+++ costcook/src/main/java/com/costcook/repository/RecipeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.costcook.entity.Recipe;
 
 @Repository
-public interface RecipeRepository extends JpaRepository<Recipe, Integer> {
+public interface RecipeRepository extends JpaRepository<Recipe, Long> {
 
 	
 }
